переносит генерацию фильтров в константы
